# APP/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/groups": {
      "post": {
        "operationId": "GroupsController_createGroup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      },
      "get": {
        "operationId": "GroupsController_listGroups",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      }
    },
    "/api/groups/admin/discovery": {
      "get": {
        "operationId": "GroupsController_listAdminDiscoverableGroups",
        "parameters": [
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "serverId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminDiscoverableGroupResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      }
    },
    "/api/groups/admin/categories": {
      "get": {
        "operationId": "GroupsController_listManageableCategories",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ManageableCategoryResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      },
      "patch": {
        "operationId": "GroupsController_renameCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenameCategoryResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      },
      "delete": {
        "operationId": "GroupsController_resetCategory",
        "parameters": [
          {
            "name": "name",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetCategoryResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      }
    },
    "/api/groups/admin/categories/archive": {
      "patch": {
        "operationId": "GroupsController_archiveCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArchiveCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArchiveCategoryResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      }
    },
    "/api/groups/{groupId}/admin/join": {
      "post": {
        "operationId": "GroupsController_adminJoinGroup",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      }
    },
    "/api/groups/{groupId}": {
      "get": {
        "operationId": "GroupsController_getGroup",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      },
      "patch": {
        "operationId": "GroupsController_updateGroup",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      }
    },
    "/api/groups/{groupId}/invite-candidates": {
      "get": {
        "operationId": "GroupsController_listInviteCandidates",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserOptionResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      }
    },
    "/api/groups/{groupId}/members": {
      "post": {
        "operationId": "GroupsController_inviteMember",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteGroupMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      }
    },
    "/api/groups/{groupId}/members/{memberUserId}/role": {
      "patch": {
        "operationId": "GroupsController_updateMemberRole",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberUserId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupMemberRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      }
    },
    "/api/groups/{groupId}/members/{memberUserId}": {
      "delete": {
        "operationId": "GroupsController_removeMember",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberUserId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      }
    },
    "/api/groups/{groupId}/leave": {
      "delete": {
        "operationId": "GroupsController_leaveGroup",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveGroupResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      }
    },
    "/api/groups/{groupId}/archive": {
      "patch": {
        "operationId": "GroupsController_archiveGroup",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArchiveGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      }
    },
    "/api/groups/{groupId}/mark-read": {
      "post": {
        "operationId": "GroupsController_markGroupRead",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkGroupReadResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      }
    },
    "/api/groups/{groupId}/read-cursor": {
      "patch": {
        "operationId": "GroupsController_advanceReadCursor",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceReadCursorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvanceReadCursorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      }
    },
...
  },
  "info": {
    "title": "SekerChat HTTP API",
    "description": "Published contract for browser, CLI, reminder, and mobile clients.",
    "version": "0.8.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "seker_access"
      }
    },
    "schemas": {
...
      "CreateGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80
          },
          "category": {
            "type": "string",
            "maxLength": 40
          },
          "serverId": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "GroupServerResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "archivedAt": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "avatarUrl",
          "archivedAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "GroupMessagePreviewResponseDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "nullable": true
          },
          "senderId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "TEXT",
              "IMAGE",
              "FILE"
            ]
          }
        },
        "required": [
          "text",
          "senderId",
          "type"
        ]
      },
      "GroupWorkStateSummaryResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "updatedAt"
        ]
      },
      "GroupArtifactConfirmationResponseDto": {
        "type": "object",
        "properties": {
          "isConfirmed": {
            "type": "boolean"
          },
          "confirmedAt": {
            "type": "string",
            "nullable": true
          },
          "confirmedByUserId": {
            "type": "string",
            "nullable": true
          },
          "confirmedByDisplayName": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "isConfirmed",
          "confirmedAt",
          "confirmedByUserId",
          "confirmedByDisplayName"
        ]
      },
...
      "GroupResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "serverId": {
            "type": "string",
            "nullable": true
          },
          "server": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/GroupServerResponseDto"
              }
            ]
          },
          "isDM": {
            "type": "boolean"
          },
          "latestMessage": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/GroupMessagePreviewResponseDto"
              }
            ]
          },
          "serverAvatarUrl": {
            "type": "string",
            "nullable": true
          },
          "workState": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/GroupWorkStateSummaryResponseDto"
              }
            ]
          },
          "artifactConfirmation": {
            "$ref": "#/components/schemas/GroupArtifactConfirmationResponseDto"
          },
          "archivedAt": {
            "type": "string",
            "nullable": true
          },
          "categoryArchivedAt": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "createdById": {
            "type": "string"
          },
          "currentUserRole": {
            "enum": [
              "MEMBER",
              "ADMIN"
            ],
            "type": "string"
          },
          "unreadCount": {
            "type": "number"
          },
          "memberCount": {
            "type": "number"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupMemberResponseDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "category",
          "isDM",
          "latestMessage",
          "serverAvatarUrl",
          "workState",
          "artifactConfirmation",
          "archivedAt",
          "createdAt",
          "updatedAt",
          "createdById",
          "currentUserRole",
          "unreadCount",
          "members"
        ]
      },
      "AdminDiscoverableCreatorResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "email",
          "displayName"
        ]
      },
      "AdminDiscoverableGroupResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "serverId": {
            "type": "string",
            "nullable": true
          },
          "server": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/GroupServerResponseDto"
              }
            ]
          },
          "serverAvatarUrl": {
            "type": "string",
            "nullable": true
          },
          "archivedAt": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "createdBy": {
            "$ref": "#/components/schemas/AdminDiscoverableCreatorResponseDto"
          },
          "workState": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/GroupWorkStateSummaryResponseDto"
              }
            ]
          },
          "memberCount": {
            "type": "number"
          },
          "isCurrentUserMember": {
            "type": "boolean"
          },
          "currentUserMembershipRole": {
            "type": "string",
            "nullable": true,
            "enum": [
              "MEMBER",
              "ADMIN"
            ]
          },
          "canSelfJoin": {
            "type": "boolean"
          },
          "visibilityReason": {
            "type": "string",
            "enum": [
              "current_member",
              "archived_admin_override",
              "admin_override"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "category",
          "serverId",
          "server",
          "serverAvatarUrl",
          "archivedAt",
          "createdAt",
          "updatedAt",
          "createdBy",
          "workState",
          "memberCount",
          "isCurrentUserMember",
          "currentUserMembershipRole",
          "canSelfJoin",
          "visibilityReason"
        ]
      },
      "ManageableCategoryResponseDto": {
        "type": "object",
        "properties": {
          "serverId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "groupCount": {
            "type": "number"
          },
          "archivedGroupCount": {
            "type": "number"
          },
          "activeGroupCount": {
            "type": "number"
          },
          "latestUpdatedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "serverId",
          "name",
          "groupCount",
          "archivedGroupCount",
          "activeGroupCount",
          "latestUpdatedAt"
        ]
      },
      "ArchiveCategoryDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "object",
            "default": ""
          },
          "archive": {
            "type": "boolean"
          }
        },
        "required": [
          "category"
        ]
      },
      "ArchiveCategoryResponseDto": {
        "type": "object",
        "properties": {
          "serverId": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "archivedAt": {
            "type": "string",
            "nullable": true
          },
          "groupCount": {
            "type": "number"
          }
        },
        "required": [
          "serverId",
          "category",
          "archivedAt",
          "groupCount"
        ]
      },
      "RenameCategoryDto": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "maxLength": 40
          },
          "to": {
            "type": "string",
            "maxLength": 40
          }
        },
        "required": [
          "from",
          "to"
        ]
      },
      "RenameCategoryResponseDto": {
        "type": "object",
        "properties": {
          "serverId": {
            "type": "string"
          },
          "from": {
            "type": "string"
          },
          "to": {
            "type": "string"
          },
          "updatedGroupCount": {
            "type": "number"
          }
        },
        "required": [
          "from",
          "to",
          "updatedGroupCount"
        ]
      },
      "ResetCategoryResponseDto": {
        "type": "object",
        "properties": {
          "serverId": {
            "type": "string"
          },
          "reassignedToServerId": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "reassignedTo": {
            "type": "string"
          },
          "updatedGroupCount": {
            "type": "number"
          }
        },
        "required": [
          "category",
          "reassignedTo",
          "updatedGroupCount"
        ]
      },
      "UserOptionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "SUPER_ADMIN",
              "MEMBER",
              "ADMIN"
            ]
          }
        },
        "required": [
          "id",
          "email",
          "displayName"
        ]
      },
      "UpdateGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80
          },
          "category": {
            "type": "string",
            "maxLength": 40
          },
          "serverId": {
            "type": "string"
          }
        }
      },
      "InviteGroupMemberDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "email"
        ]
      },
      "UpdateGroupMemberRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "MEMBER",
              "ADMIN"
            ]
          }
        },
        "required": [
          "role"
        ]
      },
      "LeaveGroupResponseDto": {
        "type": "object",
        "properties": {
          "groupId": {
            "type": "string"
          },
          "left": {
            "type": "boolean"
          },
          "archivedAfterLeave": {
            "type": "boolean"
          }
        },
        "required": [
          "groupId",
          "left"
        ]
      },
      "ArchiveGroupDto": {
        "type": "object",
        "properties": {
          "archive": {
            "type": "boolean"
          }
        }
      },
      "MarkGroupReadResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ]
      },
      "AdvanceReadCursorDto": {
        "type": "object",
        "properties": {
          "eventSequence": {
            "type": "string",
            "maxLength": 19,
            "pattern": "^[1-9]\\d*$"
          }
        },
        "required": [
          "eventSequence"
        ]
      },
      "AdvanceReadCursorResponseDto": {
        "type": "object",
        "properties": {
          "groupId": {
            "type": "string"
          },
          "lastReadEventSequence": {
            "type": "string"
          },
          "changed": {
            "type": "boolean"
          }
        },
        "required": [
          "groupId",
          "lastReadEventSequence",
          "changed"
        ]
      },
...
    }
  }
}